+++ base/quaternary-tiling-block-size/lib/main.js
@@ -20,8 +20,7 @@
 
 // MODULES //
 
-var bytesPerElement = require( './../../../base/bytes-per-element' );
-var defaults = require( './defaults.js' );
+var blockSize = require( './../../../base/tiling-block-size' );
 
 
 // MAIN //
@@ -41,42 +40,7 @@ var defaults = require( './defaults.js' );
 * // returns <number>
 */
 function quaternaryBlockSize( dtypeX, dtypeY, dtypeZ, dtypeW, dtypeU ) {
-	var nbx;
-	var nby;
-	var nbz;
-	var nbw;
-	var nbu;
-	var max;
-
-	nbx = bytesPerElement( dtypeX );
-	nby = bytesPerElement( dtypeY );
-	nbz = bytesPerElement( dtypeZ );
-	nbw = bytesPerElement( dtypeW );
-	nbu = bytesPerElement( dtypeU );
-	if (
-		nbx === null ||
-		nby === null ||
-		nbz === null ||
-		nbw === null ||
-		nbu === null
-	) { // e.g., "generic" arrays
-		return defaults.BLOCK_SIZE_IN_ELEMENTS;
-	}
-	// Find the largest element size among all five arrays...
-	max = nbx;
-	if ( nby > max ) {
-		max = nby;
-	}
-	if ( nbz > max ) {
-		max = nbz;
-	}
-	if ( nbw > max ) {
-		max = nbw;
-	}
-	if ( nbu > max ) {
-		max = nbu;
-	}
-	return ( defaults.BLOCK_SIZE_IN_BYTES/max )|0; // asm type annotation
+	return blockSize( [ dtypeX, dtypeY, dtypeZ, dtypeW, dtypeU ] );
 }