+++ base/quinary-tiling-block-size/examples/index.js
@@ -20,7 +20,7 @@
 
 var dtypes = require( './../../../dtypes' );
 var cartesianPower = require( '@stdlib/array/base/cartesian-power' );
-var promotionRules = require( './../../../promotion-rules' );
+var promoteDataTypes = require( './../../../base/promote-dtypes' );
 var quinaryBlockSize = require( './../lib' );
 
 // Generate a list of input ndarray dtype quintuplets:
@@ -32,8 +32,8 @@ var b;
 var i;
 console.log( 'block_size, xdtype, ydtype, zdtype, wdtype, udtype, vdtype' );
 for ( i = 0; i < dt.length; i++ ) {
-	t = promotionRules.apply( null, dt[ i ] );
-	dt[ i ].push( ( t === -1 ) ? 'generic' : t );
+	t = promoteDataTypes( dt[ i ] );
+	dt[ i ].push( ( t === null ) ? 'generic' : t );
 	b = quinaryBlockSize.apply( null, dt[ i ] );
 	console.log( '%d, %s, %s, %s, %s, %s, %s', b, dt[i][0], dt[i][1], dt[i][2], dt[i][3], dt[i][4], dt[i][5] );
 }